# backend/tests/api/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ConnectionCreate": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "additionalProperties": true,
            "default": {},
            "title": "Config",
            "type": "object"
          },
          "connector": {
            "enum": [
              "slack",
              "gmail",
              "vapi"
            ],
            "title": "Connector",
            "type": "string"
          },
          "label": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "secrets": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "title": "Secrets",
            "type": "object",
            "writeOnly": true
          }
        },
        "required": [
          "connector",
          "label"
        ],
        "title": "ConnectionCreate",
        "type": "object"
      },
      "ConnectionOut": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "connected_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connected At"
          },
          "connector": {
            "title": "Connector",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "firm_id": {
            "format": "uuid",
            "title": "Firm Id",
            "type": "string"
          },
          "has_consent_link": {
            "default": false,
            "title": "Has Consent Link",
            "type": "boolean"
          },
          "health": {
            "enum": [
              "unknown",
              "ok",
              "degraded",
              "revoked"
            ],
            "title": "Health",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "last_inbound_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Inbound At"
          },
          "last_inbound_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Inbound Type"
          },
          "last_tested_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Tested At"
          },
          "secret_hints": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Secret Hints",
            "type": "object"
          },
          "status": {
            "enum": [
              "pending",
              "connected",
              "error"
            ],
            "title": "Status",
            "type": "string"
          },
          "test_results": {
            "additionalProperties": {
              "$ref": "#/components/schemas/TestResult"
            },
            "title": "Test Results",
            "type": "object"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "webhook_url": {
            "default": "",
            "title": "Webhook Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "firm_id",
          "connector",
          "label",
          "status",
          "config",
          "secret_hints",
          "connected_at",
          "health",
          "test_results",
          "last_tested_at",
          "last_inbound_at",
          "last_inbound_type",
          "created_at",
          "updated_at"
        ],
        "title": "ConnectionOut",
        "type": "object"
      },
      "ConnectionPatch": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "secrets": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secrets",
            "writeOnly": true
          }
        },
        "title": "ConnectionPatch",
        "type": "object"
      },
...
      "ConsentLink": {
        "properties": {
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "ConsentLink",
        "type": "object"
      },
...
      "FieldError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "code",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "InboundSetup": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "webhook_url": {
            "title": "Webhook Url",
            "type": "string"
          }
        },
        "required": [
          "webhook_url",
          "detail"
        ],
        "title": "InboundSetup",
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "TestOutcome": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "detail"
        ],
        "title": "TestOutcome",
        "type": "object"
      },
      "TestRequest": {
        "additionalProperties": false,
        "properties": {
          "input": {
            "additionalProperties": true,
            "default": {},
            "title": "Input",
            "type": "object"
          },
          "tool": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool"
          }
        },
        "title": "TestRequest",
        "type": "object"
      },
      "TestResult": {
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "at",
          "detail"
        ],
        "title": "TestResult",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Lucia API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/connections/{conn_id}": {
      "delete": {
        "operationId": "deleteConnection",
        "parameters": [
          {
            "in": "path",
            "name": "conn_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conn Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Delete a connection no mapping uses",
        "tags": [
          "internal:connections"
        ]
      },
      "get": {
        "operationId": "getConnection",
        "parameters": [
          {
            "in": "path",
            "name": "conn_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conn Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Get a connection",
        "tags": [
          "internal:connections"
        ]
      },
      "patch": {
        "operationId": "updateConnection",
        "parameters": [
          {
            "in": "path",
            "name": "conn_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conn Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Update label, non-secret config or replace secrets",
        "tags": [
          "internal:connections"
        ]
      }
    },
    "/api/v1/connections/{conn_id}/consent-link": {
      "post": {
        "operationId": "createConsentLink",
        "parameters": [
          {
            "in": "path",
            "name": "conn_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conn Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentLink"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Generate a consent/install link (invalidates earlier links)",
        "tags": [
          "internal:connections"
        ]
      }
    },
    "/api/v1/connections/{conn_id}/inbound/enable": {
      "post": {
        "operationId": "enableConnectionInbound",
        "parameters": [
          {
            "in": "path",
            "name": "conn_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conn Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboundSetup"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed"
          },
          "502": {
            "description": "The provider rejected the request"
          }
        },
        "summary": "Enable inbound events and return the webhook URL",
        "tags": [
          "internal:connections"
        ]
      }
    },
    "/api/v1/connections/{conn_id}/test": {
      "post": {
        "operationId": "testConnection",
        "parameters": [
          {
            "in": "path",
            "name": "conn_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conn Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestOutcome"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Test auth, or one tool with an explicit target",
        "tags": [
          "internal:connections"
        ]
      }
    },
...
    "/api/v1/firms/{firm_id}/connections": {
      "get": {
        "operationId": "listConnections",
        "parameters": [
          {
            "in": "path",
            "name": "firm_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Firm Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConnectionOut"
                  },
                  "title": "Response Listconnections",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "List a firm's connections (never returns secrets)",
        "tags": [
          "internal:connections"
        ]
      },
      "post": {
        "operationId": "createConnection",
        "parameters": [
          {
            "in": "path",
            "name": "firm_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Firm Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed"
          },
          "502": {
            "description": "The provider rejected the request"
          }
        },
        "summary": "Add a connection (Vapi is provisioned immediately)",
        "tags": [
          "internal:connections"
        ]
      }
    },
...
  },
...
}